+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+- bootstrap: enable default S3 server-side encryption
+
 
 
 ## [0.11.0] - 2021-06-14